Beginning work on Event Creation - starting to use Interactors
I am refactoring my architecture to make the presenter more passive. I am now handling data manipulation in Interactors. I’ve created an EventsInteractor and a ContactsInteractor to both be consumed by the EventCreationPresenter, since the presenter required disparate data from both Events and Contacts.

I am probably doing this wrong, but I’m learning. Need tor read more on this pattern, but it seems to be working for now. Although this is still a work in progress.