EventCreation: Removed a bunch of pointless logic
I’m just updating and setting the contacts list to the chips input every time a contact is returned, instead of checking it against a previous list size. Not sure why I was really doing that in the first place. Any time a contact would come back null it would break the whole logic.