refactoring accepting and rejecting event Invites
Im now using a list of map entries as my dataset for the adapter. This allows me to remove the hash map and list combination I was using previously.

Now I need to pass both eventId and eventDetail when removing an item from the list because I need to be able to find the map entry in the list by a proper map entry