HOTFIX-Adding Authentication Activity
See https://stackoverflow.com/questions/51577039/check-if-the-user-has-signed-in-for-the-first-time-or-not-using-firebase-authui

Basically, the way I have been checking if authentications are for sign ups or existing users is flakey. Now I'm adding my own auth activity and ripping out FirebaseUI. Still a WIP but fixes the major issues where users were unable to sign up sometimes.